GIT: renaming utilities to utils.py
