+++ tests/conftests.py
@@ -1,3 +1,3 @@
 import pytest
 
-from lmr_analyzer import package, stop, route, geometry, analysis, utilities
+from lmr_analyzer import package, stop, route, geometry, analysis, utils
